feat: update ManageButton
